requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
html2text>=2020.1.16
tqdm>=4.66.0
lxml>=4.9.0
